DayTabs refactor
